Initial Commit
Data Analytics project completed during the 2024 Winter academic term. Collection of files used for implementation in a Google Drive and Google Colaboratory environment.

Files paths have not been adjusted for direct recreation of project (can be adapted if needed). Uploaded files, in unadjusted form, are intended to serve as example of the author's programming capability.